Hotel_Recommendation_System/requirements.txt
pandas==2.0.3
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.3
streamlit==1.25.0
